requirements.txt
wheel
djangorestframework
django-filter
drf-yasg
djangorestframework-simplejwt
psycopg2-binary
Pillow
python-decouple
python-slugify
celery
redis
pewee
django-cors-headers


